int-brain-sbc/int_brain_sbc/requirements.txt
smbus2
pyserial